+++ docs/CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [2.3.3] - 2022-12-13
 ### Changed
 - Link unacked PS update stream to PS module on product definitions sync (OSIDB-629)
 - Increase PS component name length from 100 to 255 characters (OSIDB-635)